Rename ledgerwalletmobile to ledgerlivemobile
