+++ android/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>ledgerwalletmobile</name>
-	<comment>Project ledgerwalletmobile created by Buildship.</comment>
+	<name>ledgerlivemobile</name>
+	<comment>Project ledgerlivemobile created by Buildship.</comment>
 	<projects>
 	</projects>
 	<buildSpec>
