+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.ledgerwalletmobile",
+    package = "com.ledgerlivemobile",
 )
 
 android_resource(
     name = "res",
-    package = "com.ledgerwalletmobile",
+    package = "com.ledgerlivemobile",
     res = "src/main/res",
 )
