+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'ledgerwalletmobile'
+rootProject.name = 'ledgerlivemobile'
 include ':react-native-touch-id'
 project(':react-native-touch-id').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-touch-id/android')
 include ':react-native-splash-screen'
