+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ledgerwalletmobile",
+  "name": "ledgerlivemobile",
   "displayName": "Ledger Wallet"
 }
