+++ ios/ledgerlivemobile.xcodeproj/xcshareddata/xcschemes/ledgerlivemobile.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "ledgerwalletmobile.app"
-               BlueprintName = "ledgerwalletmobile"
-               ReferencedContainer = "container:ledgerwalletmobile.xcodeproj">
+               BuildableName = "ledgerlivemobile.app"
+               BlueprintName = "ledgerlivemobile"
+               ReferencedContainer = "container:ledgerlivemobile.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ledgerwalletmobileTests.xctest"
-               BlueprintName = "ledgerwalletmobileTests"
-               ReferencedContainer = "container:ledgerwalletmobile.xcodeproj">
+               BuildableName = "ledgerlivemobileTests.xctest"
+               BlueprintName = "ledgerlivemobileTests"
+               ReferencedContainer = "container:ledgerlivemobile.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ledgerwalletmobileTests.xctest"
-               BlueprintName = "ledgerwalletmobileTests"
-               ReferencedContainer = "container:ledgerwalletmobile.xcodeproj">
+               BuildableName = "ledgerlivemobileTests.xctest"
+               BlueprintName = "ledgerlivemobileTests"
+               ReferencedContainer = "container:ledgerlivemobile.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ledgerwalletmobile.app"
-            BlueprintName = "ledgerwalletmobile"
-            ReferencedContainer = "container:ledgerwalletmobile.xcodeproj">
+            BuildableName = "ledgerlivemobile.app"
+            BlueprintName = "ledgerlivemobile"
+            ReferencedContainer = "container:ledgerlivemobile.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ledgerwalletmobile.app"
-            BlueprintName = "ledgerwalletmobile"
-            ReferencedContainer = "container:ledgerwalletmobile.xcodeproj">
+            BuildableName = "ledgerlivemobile.app"
+            BlueprintName = "ledgerlivemobile"
+            ReferencedContainer = "container:ledgerlivemobile.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ledgerwalletmobile.app"
-            BlueprintName = "ledgerwalletmobile"
-            ReferencedContainer = "container:ledgerwalletmobile.xcodeproj">
+            BuildableName = "ledgerlivemobile.app"
+            BlueprintName = "ledgerlivemobile"
+            ReferencedContainer = "container:ledgerlivemobile.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
