+++ ios/ledgerlivemobile/AppDelegate.h
