+++ ios/ledgerlivemobile/Images.xcassets/AppIcon.appiconset/1024.png
