+++ ios/ledgerlivemobile/Images.xcassets/AppIcon.appiconset/120.png
