+++ ios/ledgerlivemobile/Images.xcassets/AppIcon.appiconset/180.png
