+++ ios/ledgerlivemobile/Images.xcassets/AppIcon.appiconset/Contents.json
