+++ ios/ledgerlivemobile/Images.xcassets/Contents.json
