+++ ios/ledgerlivemobile/Images.xcassets/Logo.imageset/512-1.png
