+++ ios/ledgerlivemobile/Images.xcassets/Logo.imageset/512.png
