+++ ios/ledgerlivemobile/Images.xcassets/Logo.imageset/Contents.json
