+++ ios/ledgerlivemobileTests/Info.plist
