+++ ios/ledgerlivemobileTests/ledgerlivemobileTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface ledgerwalletmobileTests : XCTestCase
+@interface ledgerlivemobileTests : XCTestCase
 
 @end
 
-@implementation ledgerwalletmobileTests
+@implementation ledgerlivemobileTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
