--- ios/ledgerwalletmobile.xcodeproj/xcuserdata/gre.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>ledgerwalletmobile.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>2D02E47A1E0B4A5D006451C7</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>2D02E48F1E0B4A5D006451C7</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
